[WIP] webUI stuff.. Still some more things I want to knock out before
pushing to main.

credits to @Kaysharp42 for the entire webUI stuff, It's pretty much their UI with a few small changes and a different auth system.

Mainly need to setup a headless GUI of some sort so that I can turn on the twitch login/logout through the webui, and do some more testing on things.

currently I still require the usage of at least the normal client for a single login just to get the cookies.jar file, I'm hoping to change this in the coming days.